% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direbonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{dire bonjour
Une fonction pour dire bonjour poliment et sobrement}
\usage{
dire_bonjour(prenom)
}
\arguments{
\item{prenom}{une chaine de caractere}
}
\value{
la fonction renvoie le mot de bonjour
}
\description{
dire bonjour
Une fonction pour dire bonjour poliment et sobrement
}
\examples{
dire_bonjour("bernard")


}
